Move slack_agent to its own module.
