+++ src/zentropi/contrib/agents/slack/__init__.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+# coding=utf-8
+
